Added and Modified many things, check description
Added code in DataBase class to handle all the operations related to DB. Modified code in MainController to be able to fetch taskList from DataBase class and display it in the UI. Mapped newTaskButton to addNewTask function in MainController to load addTaskView file. Modified Main to perform initial function calls for DataBase class. App is running fine. One Warning of SQL.